pyproject.toml
[tool.poetry]
name = "cv-analyser"
version = "0.1.0"
description = ""
authors = ["Samuel Sublate <samuelsublate.dev@gmail.com>"]
readme = "README.md"
packages = [
    { include = "analyser" }
]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.37.0"
pandas = "^2.2.2"
plotly = "^5.23.0"
streamlit-aggrid = "^1.0.5"
streamlit-option-menu = "^0.3.13"
tinydb = "^4.8.0"
pymupdf = "^1.24.9"
langchain = "^0.2.12"
langchain-groq = "^0.1.9"
python-dotenv = "^1.0.1"
gspread = "^6.1.2"
oauth2client = "^4.1.3"
google-api-python-client = "^2.140.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
